pr(NewsItemPage): блоки контента и комментариев разнесены по отдельным компонентам
pr(NewsItemPage): блоки контента и комментариев разнесены по отдельным компонентам
