+++ src/components/NewsContentBlock/NewsContentBlock.tsx
@@ -1,0 +1,44 @@
+import { fromUnixTime, lightFormat } from 'date-fns';
+import {
+  By,
+  ByTimeWrapper,
+  ContentBottomWrapper,
+  ContentHeaderWrapper,
+  ContentWrapper,
+  LinkToNews,
+  Points,
+  Time,
+} from './NewsContentBlock.styled.ts';
+import { NewsItemTitle } from '../../commons/NewsItemTitle.ts';
+import { useSelectorNewsItem } from '../../store/store.ts';
+
+export default function NewsContentBlock() {
+  const newsItem = useSelectorNewsItem();
+
+  const timestampToDate = (timestamp: number) => {
+    const date = fromUnixTime(timestamp);
+    return lightFormat(date, 'HH:mm dd.MM.yy').toString();
+  };
+
+  return (
+    newsItem !== null && (
+      <ContentWrapper>
+        <ContentHeaderWrapper>
+          <NewsItemTitle>{newsItem.title}</NewsItemTitle>
+          <LinkToNews>
+            Link: <a href={newsItem.url}>{newsItem.url}</a>
+          </LinkToNews>
+        </ContentHeaderWrapper>
+        <ContentBottomWrapper>
+          <Points>{newsItem.points} points</Points>
+          <ByTimeWrapper>
+            <By>
+              By: <span>{newsItem.user}</span>
+            </By>
+            <Time>{timestampToDate(newsItem.time)}</Time>
+          </ByTimeWrapper>
+        </ContentBottomWrapper>
+      </ContentWrapper>
+    )
+  );
+}
